requirements.txt
# FastAPI
fastapi==0.115.2
uvicorn==0.32.0

# Data
requests==2.32.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
pydantic[email]==2.9.2

# Config
pydantic-settings==2.6.0
python-multipart==0.0.17

# Security
bcrypt==4.2.0
PyJWT==2.9.0
